nuwa/pyproject.toml
[project]
name = "nuwa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "ollama>=0.6.1",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]
